requirements.txt
opencv-python==4.5.4.60
pandas==1.3.3
pdf2image==1.16.0
Pillow==8.4.0
poppler-utils
pytesseract==0.3.8
